Split Token into different header file
